requirements.txt
requests>=2.31.0
authlib>=1.2.0
flask>=2.3.2
flask-caching>=2.0.0
flask-limiter>=3.3.0
flask-cors>=4.0.0
pandas>=2.0.3
python-dotenv>=1.0.0
xlsxwriter>=3.1.0
openpyxl>=3.1.0
gunicorn>=21.2.0    